// src/lib/api/generated/model/notificationPageSchema.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * COMPASS API
 * The canonical backend API for COMPASS clients.
 * OpenAPI spec version: 1.0.0
 */
import type { NotificationSchema } from './notificationSchema';

export interface NotificationPageSchema {
  items: NotificationSchema[];
  page: number;
  page_size: number;
  total: number;
}